requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
Django==4.2.23
django-bootstrap-v5==1.0.11
django-debug-toolbar==5.2.0
django-jazzmin==3.0.1
django-modeltranslation==0.19.15
dotenv==0.9.9
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
python-dotenv==1.1.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.0
whitenoise==6.9.0
